requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openpyxl==3.1.2
python-dotenv==1.0.0
email-validator==2.1.0
httpx==0.25.1
pydantic==1.10.13
xlrd==2.0.1